% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_anime.R
\name{use_anime}
\alias{use_anime}
\title{use_anime}
\usage{
use_anime()
}
\description{
Use anime js in the application
}
\examples{
\dontrun{
library(shiny)
library(shiny.anime)

ui <- fluidPage(
    use_anime(),
    actionButton(
        "move",
        "Move"
    )
)

server <- function(input, output, session) {
    position <- reactiveVal(0)

    observeEvent(input$move, {
        movement <- position() + 30

        if (movement >= 200) {
            movement <- 0
        }

        position(movement)

        trigger_animation(
            targets = "#move",
            translateX = paste0(movement, "px")
        )
    })
}

shinyApp(ui = ui, server = server)
}
}
