% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_animation.R
\name{trigger_animation}
\alias{trigger_animation}
\title{trigger_animation}
\usage{
trigger_animation(session, targets, ...)
}
\arguments{
\item{session}{session}

\item{targets}{css selectors for target elements}

\item{...}{arguments passed to anime function; for details check https://animejs.com/documentation}
}
\description{
Trigger animation for specified targets
}
